+++ src/pages/dataoffering/resources/addresource/catalog/ResourceCatalogsPage.vue
@@ -1,0 +1,3 @@
+<template src="./ResourceCatalogsPage.html"></template>
+<script src="./ResourceCatalogsPage.js"></script>
+<style src="./ResourceCatalogsPage.css"></style>