+++ src/pages/dataoffering/resources/addresource/metadata/ResourceMetaDataPage.js
@@ -20,10 +20,7 @@ export default {
             languageItems: [],
             paymentMethod: "undefined",
             // TODO use enum API endpoint for payment methods
-            paymentMethods: [{ "originalName": "undefined", "displayName": "Undefined" },
-            { "originalName": "fixedPrice", "displayName": "Fixed price" },
-            { "originalName": "free", "displayName": "Free" },
-            { "originalName": "negotiationBasis", "displayName": "Negotiation basis" }],
+            paymentMethods: [],
             valid: false,
             defaultRule: validationUtils.getRequiredRule(),
             urlRule: validationUtils.getUrlRequiredRule(),
@@ -32,14 +29,19 @@ export default {
         };
     },
     mounted: function () {
-        this.loadLanguages();
+        this.loadEnumValues();
     },
     methods: {
-        async loadLanguages() {
+        async loadEnumValues() {
             try {
-                this.$data.languageItems = (await DataUtils.getLanguages());
+                this.$data.languageItems = await DataUtils.getLanguages();
             } catch (error) {
                 errorUtils.showError(error, "Get languages");
+            }
+            try {
+                this.$data.paymentMethods = await DataUtils.getPaymentMethods();
+            } catch (error) {
+                errorUtils.showError(error, "Get payment methods");
             }
         },
         nextPage() {