+++ CHANGELOG.md
@@ -2,6 +2,21 @@
 All notable changes to this project will be documented in this file.
 (Skipped major version 1, 2 and 3 to match versioning of IDS DataSpaceConnector)
 
+## [8.5.0] - 2021-09-30
+
+### Added
+- Create, update and delete app stores
+- Manage catalogs
+- Add resources to catalogs
+- Settings: Radio buttons to use/not use proxy
+- Settings: Show available DSC update
+
+### Fixes
+- Settings: Correct handling of proxy username/password (new DSC attribute to detect set authentication)
+- Data consumption: wider contract dialog (fixes truncated date fields)
+- Data offering: load resource with no route
+- Data offering: block navigation menu while saving resource
+
 ## [8.4.0] - 2021-08-20
 
 ### Added